requirements/dev.txt
# Only development-specific requirements but imports the common ones too.
-r common.txt

black == 23.3.0
django-debug-toolbar == 4.0.0
django-extensions == 3.2.1
drf-spectacular == 0.26.1
model-bakery == 1.11.0
pytest == 7.3.1
pytest-django == 4.5.2
pytest-watch == 4.2.0
